Refactor replay CLI
