+++ simulator/src/replay/error.rs
@@ -1,0 +1,7 @@
+use radix_engine_interface::prelude::ParseNetworkError;
+
+#[derive(Debug)]
+pub enum Error {
+    ParseNetworkError(ParseNetworkError),
+    IOError(std::io::Error),
+}
